docker: Ensure journal client services will be up and running
Services acting as journal clients were always exiting prematurely due
to not created kafa topics they were trying to subscribe to.

So ensure needed topics will be created when starting the kafka service
and wait for their creation before starting the journal client services.

As a consequence, this makes a call to "docker-compose up -d" command a
little bit longer before it returns.