docker: Bump compose file format version to 2.1
rDENV5db372fcb5abf45f3a95545029e525a2a149a91c introduced the use of the
healthcheck feature from docker-compose.

The healtcheck feature was introduced with version 2.1 of the Compose
file format.

So bump Compose file format version to 2.1 in compose files in order
to avoid the following error when using docker-compose < 1.27:

"depends_on contains an invalid type, it should be an array"