docker: Install scheduler runner which deals with recurrent tasks
That will take care of scheduling recurring ingestion tasks (load-git, load-hg, etc...).
Those were currently no longer automatically scheduled since the scheduler/lister
refactoring. So they needed some actions from the docker environment user.

With the following, the user won't have to do anything to do anymore as it's supposed to
be started with the remaining part of the environment.

Related to T3667