Dockerfile: Enable metadata loading
When installed swh-loader-metadata hooks itself in the loader process,
and fetches metadata for origins created from a known lister (currently
only github).

Note that by default, requests are unauthenticated, so they will be
throttled after 60 requests per hour, causing the git loader to be
stuck while waiting for the rate-limit to expire.