docker: Improve swh-scheduler-schedule-recurrent service startup
In order for the origin loading tasks resulting from a listing to be
successfully scheduled by that service, the loading task types must
have been registered in the scheduler database prior the execution of
the main command of that service.

So make the service depends on the swh-loader one and use a Python script
to check the celery worker used to execute loading tasks have been started.