docker: deploy replica and mirror behind dedicated nginx proxies
Deploying them behind the main proxy using SCRIPT_NAME wsgi env var
kind-of work, but not for static files and a few other assets, so it's
easier that way.

mirror is hosted on :5081 and replica is hosted on :5082.

Also deploy them a "mirror" so there is a visible logo showing it's not
the main instance.