docker: add a test for the mirror
This also add a new service (kafka-rest) in the main compose file to
provide a (Confluent) Kafka REST proxy to make it easier to query the
actual state of kafka (topics, consumers, etc.) from tests or
entrypoints.

This also come with a number of changes in the conftest to add support
for multiple compose files to be used in a test session as well as
several possible endpoints to call the SWH public API from. This
requires all the session fixtures to be changed into module scoped ones.