docker: fix swh-graph and make it use a named volume
The $DATADIR/g/edges/ must be created in the entrypoint otherwise the
compressed graph creation will fail.

The creation of the /srv/softwareheritage/graph directory in the
Dockefile is needed so the named volumes mount point exists with proper
file permissions (otherwise it's owned by root:root).

And remove mistakenly added src volumes...