docker/tests: Allow to run test sessions in a container and in parallel
These changes enable to run the docker tests in a container by using
the docker bridge network gateway IP instead of the localhost domain
name in the URLs that are requested during tests execution.

They also enable to execute multiple test sessions in parallel by patching
the compose files at the beginning of each session to ensure service ports
are bound to free ports on the docker host.