docker: rework the scrubber and add a simple test for it
The swh-scrubber service is now only dedicated to initializing the
scrubber DB then wait forever so it can be used to exec check commands.

Note that it is NOT started by default (using `replicas=0`) to let
the user/test start the actual scrubbing session when ready to do so
(aka when there have been content ingested in the archive).

Other changes:
- Replace usage of `python3 -m swh` by calls to the `swh` cli tool (for
  some reason the former does not work any more here, plus the later form
  is used in other places.)

- Replace the custom DB init code by the `swh_setup_db` helper function,
  and make sure the service stays up after the scrubbing is done
  (sleeping) so that the `compose up --wait` command does not fails if
  this scrubbing is completed too fast.

The added test is rather basic for now: it creates a scrubbing session
for snapshots with 4 partitions, then run it and checks for the status
of the scrubbing session. It lacks a lot of tested features, like
ingesting corrupted objects, having holes, etc.