Stop exposing hardcoded ports from services on the docker host
Only expose ports for nginx services using compose's automatic host port
selection. As a consequence, there is no need for patching compose files
any more to run tests.

Also make nginx services to use 80 as internal port (there is no need
for using 508x anywhere).

This requires a small hack in test_vault when using azure as backend,
because the azure service is expected to be accessible from both the
host (where tests are running) and the compose services with the same
url, which is not easily feasible when running azurite in the compose
environment. The issue is, when using azure as backend for vault, it
generates a 301 response with the url of the asset stored in azure as
Location header, but in this context, the URL cannot be resoved from
the host.

It used to be handled by patching the configuration and compose files on
the fly when running tests. The proposed approach is a bit less invasive
(but still not very pretty).