docker: Make loader's storage pipeline config more robust
Replace the retry proxy by a tenacious one and move the filter one at
the entry of the pipeline.

It should mitigate spurious hash collision errors when loading several
origins in parallel (which happen when there are more than one origin
to load, now that the origins() fixture scales the swh-loader service
up if needed).