Test removal/restore from objstorage with swh-alter
We now test that `swh alter remove` actually removes content from the
primary objstorage and from extra objstorages. We also test that `swh
alter recovery-bundle restore` actually restores content to the
objstorage.

To do this, we need to compute hashes (only SHA1 for now) of the content
bytes. We extract these bytes from the recovery bundle for
convenience, but this is a fairly slow operation. Therefore we only take
a random subset (currently 5) of the removed Content objects in order
to reduce the tests runtime.

In order to test extra objstorages, we need to also spin a content
replayer. We adapt what was made for the mirror to do so.

Depends swh-alter!15