Add a docker-helper service
This service allows to retrieve some context information from the
running compose environment.
For now, only served public IP ports for a running service is available.
This allows to adapt the configuration of services that need to
generate public facing URLs (like azurite with vault or keycloak).

Use this service to configure the frontend URL at startup in keycloak.