tox: Set suicide_timeout option to 6O seconds
It enables tox to wait before sending the signal brutally killing foreground
processes, thus giving time to pytest to execute the teardown fixtures.

Previously, started containers were leaked when hitting Ctrl-C as tox
was running because pytest was killed before it could execute the code
stopping containers.

This change should also ensure containers are not leaked on Jenkins when the
docker-tests job is manually aborted by an user.