Disable Elasticsearch “disk allocation decider”
Having 18 GiB free on a disk of 425 GiB should be fine to run our tests.
But this is 97% full and above the “high watermark”, so the disk
allocation decider would lock any writes from happening to
Elasticsearch. Let’s just disable the whole mechanism, we do
few writes and only have a single shard.