Improve tests for swh-alter
The semantic of the tests assumed that swh-alter had been foked from
swh-py-template. This was not true as files were copied instead. We now
use the main `swh-py-template` repository and `lunar/swh-py-template`.

We now ensure–by checking removed swhids and swhids referenced by the
removal set-that the “Marking candidates used elsewhere” step of the
selection algorithm work. This would not have been the case without
the `record_references` storage proxy configured in c6cf71e38e.

The objstorage tests that were done on the fork now has to be done on
the initial repository to ensure a sample of Content objects big enough
to matter.

We also test what happens when we remove all origins at the same time.