tests: Add test for bulk on-demand archival feature
Add a docker test for the new bulk on-demand archival feature
enabling a trusted user to submit a list of origins to load
into the archive through a single Web API request.

The test checks all steps of the save bulk pipeline are correctly
executed once a save bulk request is received.

Related to product-management/core-platform#11.