Also add a public view to the mirror setup
In e5de22be8a8, we have introduced new containters to differentiate
when services need access to the full archive or if they only need
the view aimed at the general public. The latter has be to be filtered
by the masking proxy.

In the same way as before, the container `swh-mirror-storage-public` is exposed
from Nginx at `/rpc/storage-public/`. `swh-mirror-vault` and
`swh-mirror-web` now use these new endpoints.