tests/mirror: make tests for removal notification handling pass all at once
They used to work individually but not when executed all at once, because of
the stateful nature of the masking proxy and the mirror storage (these are
not reset between tests).

Rework the fixtures to make them let the masking and storage DB in the
expected state for any other test.

Also define the mains 'swh-alter' service in the compose.mirror.yml file
instead of using the compose.alter.yml to prevent spawning a cassandra
setup for these tests, improving significantly their execution time.