tests: Use a single compose session to run most of the tests
By default, the compose session is restarted from scratch when processing
a test module.

But most of the tests can be executed using a single session without
interfering with each other so better proceeding like this to avoid
wasting time when fully reloading all compose services for each test
module.

Tests modules that does not require a full reset of the compose session
can override the new reset_compose_session fixture so that it returns False.

Use that new feature in most of the tests as it greatly speedups the overall
execution of tests, from more than 30 minutes locally to less than 20 minutes.

Some tests have been adapted accordingly to avoid their failures.